updates to experiments

+++ bc.py
@@ -21,7 +21,7 @@ class SpaceInvLearner(nn.Module):
 
     def forward(self, x):
         x = x.view(x.size(0), -1)
-        x = F.relu(self.fc1(x))
+        x = F.relu(self.fc1(x.float()))
         x = F.relu(self.fc2(x))
         out = F.tanh(self.fc_out(x))
         return out
